Separate vnode & attrs compat into render module (-5 B)
